Add files via upload
This is the ECM K to 2 pion decay issue. And momentums etc. 
The event number is correlated with the ecm issue. All rows with the same event ID share the issue in disjoint sets. So all rows with event number #12345 would share
the ecm conservation issue. And all with #900000 might as well. But all others did not. 

In total 5.68% or 2043 of the rows out of the total amount of rows in the data file displayed this issue. Some rows corresponded to a shared event number, usually with at most 4 entries corresponding to the same number. 

A look for other correlators, as well as a plot is underway.